Completes database and authentication
In the modals file, create a connection to the database, by using login parameters extracted from the database url, also create sql queries that create two tables having a relationship of a user id.

In the auth file, create a class with methods that register a user and also logs in a user.

Create a config file with settings for different enviroments.

create a flask application in the run.py file, inclue here also the jwt
secret key and initialise the app with jwt manager.

In the user file create endpoints that login and signup the users, and in the login endpoint, create an access token that will be given to the user after successfull login.

update requirements file

[finishes #162042673] set up project
[finishes #162047507] create register endpoint
[finishes #162048297] create endpoint that logs in a user