Creates endpoint to make a parcel delivery order
In the user views write an endpoint that creates a parcel delivery order.
Also add placeholders for all the other routes that will have effect on a parcel delivery orders eg view all parcel delivery order.

Add method to the models file that executes a sql query to store
a parcel delivery order in a database.

In the admin module create placeholders for the routes that are to be used by an admin.

[finishes #162125835] finished endpoint