[finishes #162155679] Create endpoint that cancels a delivery order
write an endpoint that changes the status of endpoint from pending to cancelled, if the parcel doesn't exist then send a message that it doesn't.

write a database query that checks if a parcel exist of a unique id and user
and update the value in the status column from pending to cancelled.